requirements.txt
xmltodict==1.0.2
python-dotenv==1.1.1
pytest==8.3.3
beautifulsoup4==4.12.2
Django==5.2.8
psycopg2-binary
alembic==1.17.2
asgiref==3.11.0
colorama==0.4.6
coverage==7.13.1
django-appconf==1.2.0
django-cleanup==9.0.0
django-imagekit==6.0.0
greenlet==3.2.4
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pilkit==3.0
pillow==12.0.0
pluggy==1.6.0
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2